requirements.txt
django 
djangorestframework
python-dotenv
langchain
langchain-community
langchain-groq